Fix RunningSession#create, change runningsession db

+++ db/migrate/20210304160459_create_running_sessions.rb
@@ -2,10 +2,10 @@ class CreateRunningSessions < ActiveRecord::Migration[6.1]
   def change
     create_table :running_sessions do |t|
       t.references :user, null: false, foreign_key: true
-      t.integer :distance
+      t.float :distance
       t.bigint :start_time
       t.bigint :finish_time
-      t.integer :duration
+      t.float :duration
       t.float :avg_pace
       t.float :avg_speed
       t.timestamps
